+++ src/test/DurationAdapter.java
@@ -1,0 +1,18 @@
+package test;
+
+import com.google.gson.*;
+import java.lang.reflect.Type;
+import java.time.Duration;
+
+public class DurationAdapter implements JsonSerializer<Duration>, JsonDeserializer<Duration> {
+
+    @Override
+    public JsonElement serialize(Duration src, Type typeOfSrc, JsonSerializationContext context) {
+        return new JsonPrimitive(src.toString());
+    }
+
+    @Override
+    public Duration deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+        return Duration.parse(json.getAsString());
+    }
+}
